refactor(monitoring): split monitoring.rs 265\u2192214 + live 67 (cycle 46)
