+++ src/bin/email_api_dir/monitoring_handlers/mod.rs
@@ -2,10 +2,12 @@
 // Public API preserved via `pub use` re-exports; parent uses `pub use monitoring_handlers::*`.
 
 mod monitoring;
+mod monitoring_live;
 mod security;
 mod shared;
 
 pub(crate) use monitoring::*;
+pub(crate) use monitoring_live::*;
 pub(crate) use security::*;
 #[allow(unused_imports)]
 pub(crate) use shared::*;
